Abstracts thrice duplicated method getIdentifierAndFileNames to a utility file and imports it into each required file.
