+++ lib/main/lang/fixmyts/quickFixes/addImportStatement.ts
@@ -3,35 +3,7 @@ import * as ast from "../astUtils";
 import {EOL } from "os";
 var { displayPartsToString, typeToDisplayParts } = ts;
 import path = require('path');
-import {Project} from "../../core/project";
-
-import {getPathCompletions} from "../../modules/getPathCompletions";
-
-function getIdentifierAndFileNames(error: ts.Diagnostic, project: Project) {
-
-    var errorText: string = <any>error.messageText;
-
-    // We don't support error chains yet
-    if (typeof errorText !== 'string') {
-        return undefined;
-    };
-
-    var match = errorText.match(/Cannot find name \'(\w+)\'./);
-
-    // If for whatever reason the error message doesn't match
-    if (!match) return;
-
-    var [, identifierName] = match;
-    var {files} = getPathCompletions({
-        project,
-        filePath: error.file.fileName,
-        prefix: identifierName,
-        includeExternalModules: false
-    });
-    var file = files.length > 0 ? files[0].relativePath : undefined;
-    var basename = files.length > 0 ? files[0].name : undefined;
-    return { identifierName, file, basename };
-}
+import {getIdentifierAndFileNames} from "../getIdentifierUtil";
 
 export class AddImportStatement implements QuickFix {
     key = AddImportStatement.name;
