+++ lib/tsconfig.json
@@ -70,6 +70,7 @@
         "./main/lang/core/languageServiceHost2.ts",
         "./main/lang/core/project.ts",
         "./main/lang/fixmyts/astUtils.ts",
+        "./main/lang/fixmyts/getIdentifierUtil.ts",
         "./main/lang/fixmyts/quickFix.ts",
         "./main/lang/fixmyts/quickFixes/addClassMember.ts",
         "./main/lang/fixmyts/quickFixes/addClassMethod.ts",
